Refactor codemirror loading of mode to shared util
Refactor cloud and local space to use the same shared instance of modified codemirror
