+++ src/cloud/lib/export.ts
@@ -11,10 +11,10 @@ import rehypeKatex from 'rehype-katex'
 import { mergeDeepRight, prop } from 'ramda'
 import gh from 'hast-util-sanitize/lib/github.json'
 import { downloadString } from './download'
-import rehypeCodeMirror from './rehypeCodeMirror'
+import rehypeCodeMirror from '../../shared/lib/codemirror/rehypeCodeMirror'
 import { SerializedDoc } from '../interfaces/db/doc'
 import { filenamify } from './utils/string'
-import { UserSettings } from './stores/settings/types'
+import { UserSettings } from './stores/settings'
 import { getGlobalCss } from '../components/GlobalStyle'
 import { selectTheme } from './styled'
 import { boostHubBaseUrl } from './consts'
