+++ src/lib/exports.ts
@@ -8,13 +8,13 @@ import rehypeStringify from 'rehype-stringify'
 import rehypeKatex from 'rehype-katex'
 import { mergeDeepRight } from 'ramda'
 import gh from 'hast-util-sanitize/lib/github.json'
-import { rehypeCodeMirror } from '../components/atoms/MarkdownPreviewer'
+import rehypeCodeMirror from '../shared/lib/codemirror/rehypeCodeMirror'
 import { Attachment, NoteDoc, ObjectMap } from './db/types'
 import { filenamify } from './string'
 import React from 'react'
 import remarkEmoji from 'remark-emoji'
 import rehypeReact from 'rehype-react'
-import CodeFence from '../components/atoms/markdown/CodeFence'
+import CodeFence from '../shared/components/atoms/markdown/CodeFence'
 import yaml from 'yaml'
 import {
   convertHtmlStringToPdfBuffer,
